+++ src/components/task-board/task-sheet/RemovePropertyButton.tsx
