+++ src/components/task-board/task-sheet/TaskSheetAssignee.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import LabelBadge from "./label-badge";
+import LabelBadge from "./LabelBadge";
 import { User } from "lucide-react";
 import TaskAssignee from "../task-card/task-assignee";
-import RemovePropertyButton from "./remove-property-button";
+import RemovePropertyButton from "./RemovePropertyButton";
 import { useTaskStore } from "@/store/useTaskStore";
 
 const TaskSheetAssignee = ({ taskId }: { taskId: string }) => {
